requirements.txt
streamlit
pillow
requests
openai
pandas
supabase